+++ crates/native/src/api.rs
@@ -2425,11 +2425,13 @@ pub enum VideoCodec {
     /// [H.264] Advanced Video Coding (AVC).
     ///
     /// [H.264]: https://en.wikipedia.org/wiki/Advanced_Video_Coding
+    #[allow(dead_code)]
     H264,
 
     /// [H.265] High Efficiency Video Coding (HEVC).
     ///
     /// [H.265]: https://en.wikipedia.org/wiki/High_Efficiency_Video_Coding
+    #[allow(dead_code)]
     H265,
 
     /// [VP8] codec.
@@ -2466,15 +2468,7 @@ pub fn video_encoders() -> Vec<VideoCodecInfo> {
         },
         VideoCodecInfo {
             is_hardware_accelerated: false,
-            codec: VideoCodec::H264,
-        },
-        VideoCodecInfo {
-            is_hardware_accelerated: false,
             codec: VideoCodec::AV1,
-        },
-        VideoCodecInfo {
-            is_hardware_accelerated: false,
-            codec: VideoCodec::H265,
         },
     ]
 }
@@ -2493,15 +2487,7 @@ pub fn video_decoders() -> Vec<VideoCodecInfo> {
         },
         VideoCodecInfo {
             is_hardware_accelerated: false,
-            codec: VideoCodec::H264,
-        },
-        VideoCodecInfo {
-            is_hardware_accelerated: false,
             codec: VideoCodec::AV1,
-        },
-        VideoCodecInfo {
-            is_hardware_accelerated: false,
-            codec: VideoCodec::H265,
         },
     ]
 }